requirements.txt
playwright==1.50.0
pytest-playwright==0.7.0
python-dotenv==1.0.1
pytest==8.3.4
